requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
certifi==2017.7.27.1
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
Django==1.10.6
django-cors-headers==2.3.0
djangorestframework==3.8.2
Fabric3==1.14.post1
httpie==0.9.9
idna==2.6
Markdown==2.6.9
paramiko==2.4.1
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.1
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
